Use UserError in case of missing input
Using UserError (previously we had a ValidationError) if
the date is not specified for a stock movement report.

Such error type is consistent with what we use in case
locations are not specified for a stock summary report
and it's a more appropriate (warning) for the user
instead of a generic server error.