Story #589 (Task #624) - Stabilise the JS unit tests.
Some of the odoo JS unit tests broke after Story #519.  The reason they broke is that the changes made weren't coded defensively enough.  Therefore, ensure that we check that when accessing attributes that the object that we are referencing is actually defined (it won't be defined in odoos core suite of tests)