[FIX] Correcting field usage for dependencies
Previously had code in udes_core that was explicitly using a
field defined in warehouse_config, but udes_core does not have a
dependency on warehouse_config.
As a result the code has now been split so as to not create undeclared
dependencies, and the tests have gone into udes_api, which requires
both udes_core and warehouse_config to work.