[IMP] udes_purchase: Add custom email and report template for reorder point
We don't require most of the infomation in the standerd attachment.
Also the correct layout was not being used.

Added new email and report, also added context to send_mail call.
Functions for both the reference and context were added to allow
them to be overridden if changes are required.

User-story: story/977