[IMP] udes_stock: Make move line key a computed field, reassign responsiblities
Move line key is now a computed field on move.line, but computation is
explicitly re-run before grouping so the value is up-to-date guaranteed
to be up to date, otherwise it would have to depend on every field in
move.line

The stock.move object is now responsible for post-reservation splitting
as that is the object being considered.

The stock.picking object is now responsible for the creation of a new
picking for split moves.

Signed-off-by: Charlie Wheeler-Robinson <crwr@pigotts.org.uk>