[FIX] udes_stock: Do not use translatable strings in xpath expressions
XPath expressions must not use human-readable text strings to locate
elements to be modified since these are likely to be changed.  In
particular, the expression will fail if it refers to a text string
subject to language translation.

Fix by matching on the "name" attribute value (and absence of the
"context" attribute) instead of matching on the element text contents.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>