[FIX] Make dest loc really optional for batch::update
Here we avoid raising a ValidationError while handling
a batch::update request in case the destination location
is not specified.

To do that, we only update the related move lines in case
the dest loc is specified.

Note that this change is necessary for implementing the
logic to handle 'dont_scan' picking in the front end,
for which the user does not scan the destination location.