[IMP] (1525) Improve suggested locations (no product case)
With this patch we improve the backend logic of the
picking::suggested_locations' "by_product" and "by_package"
policy; in case there's is no product in the drop zone,
the empty locations will be suggested.

Previously, in such case, no suggested was returned in the
endpoint response and, as a consequence, the frontend
could not discriminate between the case where empty locations
are available and when all locations are used by other
products.

Adding new unit tests for suggested location policies