[IMP] udes_stock: Post-reservation split handles non-default locations.
In order to work with moves for which the src and destination location
do not match the defaults of the picking type, we cannot just create a
new picking with the default src and dest locations of the picking type.
Instead they have to be taken from the moves that will be added to the
created picking, and for that to be possible, the moves have to be for
the same locations.

This change enforces that moves have the same src and dest locations
when applying the post reservation split, so that those locations can be
used when creating the picking.

Update Dockerfile to include new dependency EDI

Needed for story/1775 so that Putaway can be used to putaway to
somewhere outside Stock.

Story: 1775
Signed-off-by: Charlie Wheeler-Robinson <crwr@pigotts.org.uk>