[IMP] udes_stock: Add stock.picking sequence field
Provide the ability to specify an arbitrary ordering amongst
stock.picking records with the same priority and scheduled date, by
adding a standard "sequence" integer field and redefining the default
sort order.

A standard handle widget is provided in the stock.picking tree view.
Note that the sequence is (deliberately) not the highest precedence
field within the sort order, and so using the widget to attempt to
reorder stock.picking records that do not already have the same
priority and scheduled date will not work as expected.

There seems to be no sensible way to modify the behaviour of sequence
handle widgets to e.g. limit the positions in which the dragged row
can be dropped.

On the server side, we can intercept the /web/dataset/resequence
endpoint and verify that the resequencing operation has left the
records in an order that reflects the user's intention.  The
resequencing transaction can be aborted and an error message
propagated to the user, but there seems to be no way to prevent the
BasicModel.resequence() method in the UI from updating the displayed
sort order as though the operation had succeeded.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>