[IMP] udes_stock: Ability to have generic pick refactoring policies.
Add fields on stock.picking.type to hold configuration for actions
to be taken when moves of that type reach stages in their lifecycle.

This is to allow picks to be split or grouped by various criteria at
the various stages in their lifecycles to a form that best matches the
operational requirements.

There is by default a by_move_line_key method that groups by an
attribute of move lines, and facility to add custom methods as well.

Story: 1942
Signed-off-by: Charlie Wheeler-Robinson <crwr@pigotts.org.uk>