[FIX] How batch task includes packaging info
Previoulsy, the udes_sale_stock specialization of the
StockMoveLine::_prepare_task_info method was assuming
a single move line when retrieving the package info of
the sale order.

With this patch with get rid of this false assumption
by traversing to the sale order info via the move line
moves.