[TEST] Install udes-open modules during prerequisite module installation
The current Dockerfile explicitly specifies the module dependencies to
be installed.  This means effectively maintaining the dependency list
in multiple places, since the Dockerfile must be kept up to date with
any changes in any of the __manifest__.py files.

Avoid this problem (at the cost of a marginally longer test runtime)
by installing the udes-open modules themselves.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>