[FIX] udes_stock: suggeseted_location mulitple picks
Previously, the picking::suggested_locations
endpoint did assume a single picking for the
specified move lines.
As we now implement different outbound workflows
that involve the manual creation of batches,
such assumption cannot be enforced anymore.

With this patch we use the move lines to
determine the picking instances that will in
turn invoke the method to determine the
suggested locations.

This is a partial change for that, as the
optimal solution will include modifying the
API specs as we don't need the picking type to
be specified anymore.

User-story: 2163