[FIX] udes_stock: Fix new_picking_for_group()
In some cases the function was returning a log error like:
Comparing oranges and apples.
The issue was that when there is already an existing picking that
matches for the given group, some fields are checked to see if
their current value is consistent with the new one to update.
If it is not then the value of the picking is set to False to avoid
misleading information. When this happens the values sometimes are
the id of the res.partner, or another recordset. In that case we
were ending up comparing a recordset with an integer.
Fixed to get the id of the recordset instead.

Signed-off-by: Miquel PS <miquel.palahi.sitges@unipart.io>