[IMP] udes_security: Enabling auth_brute_force module with settings
Adding auth_brute_force (from server_auth) module as a udes_security
dependency. Adding default configs for the modules inbuilt settings
and adding an additional setting to allow us to disable the global
IP ban for all users - allowing this to be used with Odoo when
x_forwarded_for headers are not available when behind a proxy.