[IMP] udes_stock: improve move_line::suggested_locations
Here we add more details to the response of the
move_line::suggested_locations endpoint so that it
may indicate multiple lists of suggested locations,
each one with a different title (see specs). This
is required in order to discriminate the empty
locations from those suggested by the configured
policy.

User-story: 2521