[FIX] udes_sale_stock: Bypass PII view logging for the superuser
Logging PII accesses for the superuser adds substantial amounts of log
noise and increase the time (and therefore cost) required to diagnose
almost any problem.

The superuser may be assumed to have full access to the system,
including the ability to tamper with the log files.  In addition,
security policy dictates that administrators log in as individual
accounts with administrative permissions, rather than logging in as
the superuser account.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>