[FIX] udes_stock: Do not attempt to cancel empty pickings
Pick factorisation may result in some empty pickings.  The current
code attempts to cancel these by directly writing the "state" field to
the value "cancel", with a comment noting that action_cancel() does
not work on a picking with no associated stock moves.

The reason that action_cancel() does not work on a picking with no
associated stock moves is that the "state" field is a readonly field
computed from the associated stock move lines.  A picking with no
associated stock move lines is, by definition, in state "draft".

The current code seems to succeed in setting the state to "cancel" due
only to the permissiveness of the ORM, which (for performance reasons)
generally trusts code not to attempt undefined operations such as
writing to a non-invertible but stored computed field.

Fix by adding an "active" field to stock.picking and deactivating
empty pickings instead of cancelling them, since there is no valid way
for a picking with no moves to exist in a cancelled state.

Note that unlinking the empty pickings can causes widespread problems
since it may unexpectedly invalidate a stock.picking "self" recordset
used by any superclass methods.

Remove the now spurious stock.picking.unlink_empty() method, and
remove any methods where the only purpose was to hook in a call to
unlink_empty().

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>