[IMP] udes_stock: Improve evaluation speed of move/move_line access rules
The improvements come becuase of two changes:
The first is removing the or branch in the domain a db query is removed.
The second is done by storing the picking type directly on the move_line
which removes an unbound query on the moves.
Credit goes Miquel who came up with the fix.

Signed-off-by: Michele Costa <michele.costa@unipart.io>