[FIX] udes_stock: Fix package reservation for multiple pickings
Fix a bug when action_assign is called on more than one picking
at a time, with package reservation turned on.
The bug was caused by erroneous use of self while in a loop over
the records in self, causing all packages in all pickings in self
to have to be in the current item of the loop.

This fixes the issue and adds a unit test.

Story: 3453

Signed-off-by: Charlie Wheeler-Robinson <crwr@pigotts.org.uk>