[FIX] udes_stock, udes_sale_stock: Correct dependency and config
udes_stock had an incorrect dependency on sale_stock, in order
to set a config option. This was incorrect and causes 15 extra
modules to be installed on the mvp system, significantly increasing
risk and cost of maintainance.

This change moves the dependency and config into the udes_sale_stock
module, which houses our other work that depends on sale_stock.

Story: 3390

Signed-off-by: Charlie Wheeler-Robinson <crwr@pigotts.org.uk>