[IMP] udes_stock: Revise stock.picking.u_location_category_id dependency
Remove dependency to location.u_location_category_id from
stock.picking.u_location_category_id to avoid conflicts.
Once it is set, it only changes if the move_lines or their
locations change.
This prevents to update it when pickings are done or serialisation
erros because they are in use.

Task: 3839

Signed-off-by: Miquel PS <miquel.palahi.sitges@unipart.io>