[IMP] udes_stock: Concurrent stock moves of loose product are allowed
Multiple stock moves of the same product at the same location (i.e. quant) are allowed to exist simultaneously by bypassing the check on quant reservation constraints. Bug found when cancelling a picking assigned to a batch which results in a cancelled pick having stock reserved against it. Comment added for reference before implementation of a full solution at a later date.

Story/3762

Signed-off-by: Adam Patrick <adam.patrick@unipart.io>