[WIP] udes_stock: Added hook for custom handling of confirmed not ready pickings. Refining of batch compute state
Stock picking batch compute state is not called recursively. _remove_unready_pickings only determines which pickings to remove from the batch i.e. does not try to reserve stock. Added hook to allow determination of which pickings to retain in the batch e.g. if pickings should be retained because there is stock available to them. Default behaviour is to remove unready pickings from the batch regardless of stock availability. udes_stock unit tests updated to allow for adjusted workflow and end states of the batch.

Story/4059

Signed-off-by: Adam Patrick <adam.patrick@unipart.io>