[IMP] udes_security: Abstract only admin security check
The function that checks if the user trying to add/remove trusted user
security group is admin can be abstracted to receive the group as a
paramter, so that it can be used for other security groups.

Any attemp for an unauthorized user to change the group is now logged.

User-story: 4607

Signed-off-by: Miquel PS <miquel.palahi.sitges@unipart.io>