[ADD] udes_security: Added 'debug user' group
Added a 'debug user' group that can only be added to by the main admin user, similarly to the 'trusted user' group. Only users in the 'debug user' group can user debug mode. If a user outside of this group tries to use any form of debug mode they will be redirected to a version of the page without debug mode enabled.

Story/9774

Signed-off-by: Adam Patrick <adam.patrick@unipart.io>