[REF] udes_sale_stock: Raise batch size in cancel_sale_orders_without_availability
Due to previous improvements in the usage of the Odoo cache, there should be
enough headroom to raise the batch size in
cancel_sale_orders_without_availability by a factor of ten.
This change reduces repeated overheads in the caching of fields and bears a
small performance improvement.

Task: 10006
User-story: 9978

Signed-off-by: Sean Quah <sean.quah@unipart.io>