[FIX] udes_stock: Handle newID at stock.picking.batch computed fields
Priority and picking_type_ids are computed from the pickings inside
the batch. A change was made in a previous commit to preserve their
value in case all the pickings of the batch were removed from it.
When the id of the batch is newID it shouldn't try to preserve it.

User-story: 10023

Signed-off-by: Miquel PS <miquel.palahi.sitges@unipart.io>