[ADD] suggested_locations: Add suggested_locations module
Have a separate module that contains the different policies employed in a warehouse on
how the system determines suitable locations on where stock can be dropped.
The policies are written as usual python classes, and abstract methods are used
to ensure that the correct methods are implemented.
There are two policies used here, by product and match move line.
The preprocessing flag is used to govern where the validate locations code is
triggered.

User-story: 9944

Signed-off-by: Pete Pratt peter.pratt@unipart.io