[IMP] udes_stock: TestGoodsOut rule action set to pull_push and locations switched
TestGoodsOut had the location_id and location_src_id fields backwards when compared to the same record in UDES 11. This was likely due to the fact Stock Location Path was changed to Stock Rule, with some field names being changed. This coupled with action being set to pull only prevented the route from triggering a goods-out picking being created.

Story/9859

Signed-off-by: Peter Clark <peter.clark@unipart.io>