[FIX] udes_stock: Added check for system user in get_user_warehouse
A check is carried out in get_user_warehouse to see if the user is active. However the system user is excluded from searches. An earlier check is now carried out to see if the user_id matches SUPERUSER_ID. If it doesn't then the check is carried out as before.

Story/9859

Signed-off-by: Peter Clark <peter.clark@unipart.io>