[IMP] udes_stock: Add drop off support to update_picking
When a destination location is passed to update_picking and no source package
name or products are specified, update the destination location of all done
move lines. This models a drop off operation.

The set of move lines to be updated may be narrowed down by passing a result
package name / pallet name.

User-story: 10335

Signed-off-by: Sean Quah <sean.quah@unipart.io>