[IMP] udes_sale_stock: Customer Reference and Requested Date fields on Sales Orders now required
Customer Reference and Requested Date fields also moved to the top section of Sales Orders view to be more prominent.

"(copy)" appended to Customer Reference when an Order is copied, as this field now needs to be set on a copied Order.

Updated create_sale function in BaseSaleUDES unit test class to now put in default Customer Reference and Requested Date.

Story/10507

Signed-off-by: Peter Clark <peter.clark@unipart.io>