[IMP] udes_stock: Allow reservation of one pallet per picking.
User-story/10340

Signed-off-by: Kevin Dwyer <kevin.dwyer@unipart.io>

Matrix picking entails a picker picking for multiple pickings within a
single batch, with picks for each picking being collected in a dedicated
pallet.

This patch adds two flags to the picking type model:
- `u_reserve_pallet_per_picking`
- `u_max_reservable_pallets`
to enable matrix picking and to control the number of pallets a picker
may reserve.

This patch adds one field
- `u_reserved_pallet`
on the picking model to track pallet reservations, and the necessary
bookkeeping code.