[IMP] udes_stock: _get_inventory_lines_values refactored
Refactored _get_inventory_lines_values from core Odoo, with logic for filtering and retrieving line values split out into separate functions which can be easily overridden in other modules if needed.

Rather than using SQL to retreive the values to use for Inventory Lines, a search with the appropriate filter is now carried out on Quants, with a key used to identify unique quants in a dictionary of values. If the key is already found, the quantity is updated.

Story/11983

Signed-off-by: Peter Clark <peter.clark@unipart.io>