[ADD] udes_stock: Line Updated computed flag field added to Inventory Line
Added computed Line Updated field which is used to determine whether the line should be highlighted on the Inventory Adjustment screen.

The computed field allows for other custom modules to easily override the checks required to determine if the line should be highlighted as having been updated, rather than having to completely override the expression in the view.

Story/12266

Signed-off-by: Peter Clark <peter.clark@unipart.io>