[ADD] udes_sale_stock: Make the cancelled checkbox on sale order lines work
Sale order lines may be manually cancelled if there are no pickings in
progress. Note that this is not the same as the outbound process being
in progress: sale order lines may be cancelled midway through the outbound
process as long as no one is working on a picking related to it.

Sale order lines may not be cancelled once their products have progressed
to a certain stage of the outbound process, controlled by the
u_disallow_manual_sale_order_line_cancellation_at_picking_type_ids setting
on the warehouse. Eg. cancellation can be disallowed once products are about
to be dispatched. However once those products have left the warehouse and
are considered delivered, the remaining quantity in the sale order line
may be cancelled again.

User-story: 13237

Signed-off-by: Sean Quah <sean.quah@unipart.io>