[IMP] udes-open: Create pickings unlocked
story/10434

Signed-off-by: Kevin Dwyer <kevin.dwyer@unipart.io>

Pickings are created as locked by default, but this means that once a
draft picking has been saved and exited in the desktop UI, it is no
longer possible to add moves.

There's no benefit in locking draft pickings in this way, so create them
as unlocked by default.