[IMP] udes_stock: Option to unreserve partials in check availability
story/14286

Signed-off-by: Kevin Dwyer <kevin.dwyer@unipart.io>

`StockPicking._reserve_stock_assign` will unreserve partially reserved
moves if `u_handle_partials` is enabled but `u_handle_partial_lines` is
not.  However this functionality is not triggered if reservation is
performed via `action_assign()` (for example, using the check
availability button.

This patch moves the unreserving function to `StockMove` and provides a
hook so that partial moves can be unreserved during `action_assign()` if
desired.