[IMP] udes_stock: Improvements for _compute_state
On databases with a large amount with locations with a high number quants the "location.quant_ids" can slow down the initial computation significantly. As we only care about whether a location has any quants, doing a search count will speed this check up significantly.

State is now only set if a location is internal or countable, otherwise it is set to False as in this case the stock status is not relevant.

Story/15029

Signed-off-by: Peter Clark <peter.clark@unipart.io>