[REF] udes_priorities: change priorities user restrictions
User-story: 20

A Priority can now be deleted or archived only by admins or trusted
users if there are no transfers with this priority assigned to them.

A stock transfers count is now added to Priorities showing how many
transfers are using the priority. This is helpful to understand how
changes could affect stock tranfers.

Transfer types can now be added to a priority at any time and deleted
only if there are no transfers of the same type using the priority.

Reference field now cannot be changed by any user, including trusted
or admin.

Picking types from a priority can now be removed only if there are
no stock transfers assigned to the priority. If the user tries to
remove a picking type that has transfers assigned then a message will
popup when trying to save. Originally a similar method was implemented
on picking_type_ids with an api.onchange decorator but Odoo was
creating further issues as a spur priority record was being created
and saved in the database.

Unit Tests have been added to confirm the changes are effective.

Signed-off-by: Lorenzo Cucurachi <lorenzo.cucurachi@unipart.io>