[IMP] udes_stock: Consider inactive variants for default_code and barcode fields
story/934

- Add context {"active_test": False} to various product.template views
which results in barcode (a related field) persisting after product
archival from the UI or EDI
- Add a new filter 'Active' and enabled it by default, to keep
the behaviour where inactive products are hidden by default.
- Override default_code compute method to include a hard domain
search of inactive variants to allow for the field to persist on the
template after de-activation via EDI.
- The compute ignores the view context and if we try to force set self
in the compute to have the context, Odoo would get scrambled and end
up setting the field value to False whenever the recomputation happened,
so this approach was not viable
- Add a one-off recomputation call for default_code to update any
existing 'bad' data to have the correct value. This is only needed
on inactive templates which have no default_code set (empty string)

Signed-off-by: Kevin Dwyer <kevin.dwyer@unipart.io>