[IMP] udes_stock: fix origin field not set correctly
The tests here use an helper method to create a picking instead
of using for example create_picking() in stock.picking model.
Bebcause of this there can be differences in the tests in terms
of how a picking is cretated which can mean a different behaviour
compared to UDES is tested.

This change aligns better how a picking is created compared to UDES.

Fixed a test that was checking the results of a method assuming that
only one picking would be created when actually a pick and putaway
picking is created.

User-story: 840

Signed-off-by: Lorenzo Cucurachi <lorenzo.cucurachi@unipart.io>