[IMP] udes_stock: Adding support for finding packages
User-story: 1169

The find_move_lines() function in stock.quant.package is not checking
for move lines that use the package as a destination package. This
change ensures this check is done.

The get_or_create() fucntion in the mixin stock model does not allow
for returning an empty result without raising an issue. Adding here
the support for such use case.

Signed-off-by: Lorenzo Cucurachi <lorenzo.cucurachi@unipart.io>