[REF] udes_get_info: Rework tests to prevent failure.
story/52

Signed-off-by: Kevin Dwyer <kevin.dwyer@unipart.io>

Some tests in test_get_info add location_id directly to
Location._get_info_field_names, and this causes the new test for
expected keys to fail, because location_id is not expected to be
present.

Removing location_id after it is added fixes the problem, but causes
other tests in test_get_info to fail.

This patch:
- ensures that location_id is removed after use in a test method
- inserts location_id as required to keep the existing tests passing

This ought to be a temporary measure, pending further investigation.