[FIX] udes_stock_picking_batch,udes_stock: Improve backorder generation
This commit:
  * `_backorder_move_lines` assumes `_requires_backorder` is called
before hand to check it is needed.
  * Adds `_check_backorder_allowed` to determine if the suggested
backorder is permitted. The restriction is you cannot move done or
cancelled moces into a backorder, and you cannot move mls into a
backorder with qty_done > 0 and retain some mls in the original picking
wiht qty_done = 0.
  * `_backorder_move_lines` can take a subset of mls and will split up
any moves that those mls point to, and remove all remaining work from
the original picking.
  * `drop_off_picked` will creata  backorder for all work not being
dropped off in the current action. It will add the picking to the same
batch if continue_batch is true and the stock is available to get more
work to do.

Signed-off-by: Pete Pratt <peter.pratt@unipart.io>