[IMP] udes_stock: Ensure that the prefixes of picking types are consistent with the UDES style and not Odoo's style.
The sequence_code field is not needed in UDES as it is used strictly in the Odoo style sequence (which we do not want). When updating picking_types, if there was sequence_code field in the values to update, Odoo was hardcoded to update the sequences on the picking types to the Odoo prefix style. This resulted in changes to prefixes when changing data on the picking type, such as updating the picking_type with data that set a sequence_code.

Additionally when the warehouse is created or updated with a new name or code, Odoo would update the picking type sequence prefixes.  _get_sequence_values() was were these new picking_type sequences prefixes were generated. This method has been overwritten to prevent this from happening.

Some unit tests were added to ensure that writing to a picking type or updating the warehouse name didn't result in a change to the picking type sequence prefix.

Story/1805

Signed off by: Jack Birch <jack.birch@unipart.io>