[IMP] udes_stock: Backorder names refer to creation order
When creating a backorder, the name folows the sequence they are created
in, and not the level they exist in the picking tree. Even if the
backorder is created from the original picking, but it is the tenth one
created, ensure that it suffixed with -0010 not -0001.
Refer to the README.md changes in this commit for more detail.

User-story: 2010

Signed-off-by: Pete Pratt <peter.pratt@unipart.io>