[ADD] udes_stock: Port location state
Port the u_is_countable, u_location_countable fields from 11 to 14.
Port the location state from 11 to 14, but rename from u_state ->
u_countable_state. The only possible location states
are:
  * empty
  * archived
  * has_stock
  * Falsey
The state is False if it is not countable.

The order of how the states are set are split into a different method so
it can be altered more easily.

User-story: 2106

Signed-off-by: Pete Pratt <peter.pratt@unipart.io>