[FIX] udes_stock_refactoring: Always set value for grouping key
When computing the grouping key for Move/Move Line,
if `compute_key` was not in context then the compute method was
previously returning without assigning a value.
In Odoo 14 this is not allowed and will lead to an exception.

To avoid this Move/Move Line will have None set by default.

Signed-off-by: Peter Clark <peter.clark@unipart.io>