[IMP] udes_stock_cron: Avoid reserving stock if insufficent
story/x1683

Signed-off-by: Kevin Dwyer <kevin.dwyer@unipart.io>

StockPicking.reserve_stock tries to allocate stock and rolls back if
there is insufficent stock to allocate.  This can be inefficient in busy
systems.  This patch changes the behaviour to check if there sufficent
stock prior to allocating, and to skip allocation if there is not.
There is a small chance that stock will be allocated in between the
check and allocation, causing a rollback, but this is acceptable.