[IMP] udes_sale_stock: Link all created pickings to the sale order. This also means all pickings will be cancelled when the order is cancelled.
Not all the moves that were being created were being associated with the sale_order_line.

To overcome this had to ensure that when the move information was passed down the chain of procurement rules, the sale_line_id was also passed down. This way each move, even down the chain of pickings, is associated with the order. This was done by default in Odoo 11 but not in Odoo 14.

Unit tests were added to check the udes_sale_stock changes.

Story/2021

Signed-off-by: Jack Birch <jack.birch@unipart.io>