[IMP] udes_stock: Hide create button in form view for specific models
Create button will no longer show in the form view for Pickings or
Inventory Adjustments.

Changes made via JS, as disabling the create action directly on the
form view also hides the dupicate action.

Longer term it would be good to have a more generic way to handle this.

Story/2466

Signed-off-by: Peter Clark <peter.clark@unipart.io>