[FIX] udes_stock_picking_batch: Always save picking_ids on batch in UI
When removing all pickings from a batch, the state will be re-computed
as done, which means that picking_ids becomes readonly. As a result
the changes are not saved, although they appear saved in the UI until
the page is refreshed or the record is re-visited.

Added force_save attribute to picking_ids to avoid this issue.

Move lines on batch also made readonly, editing these within UDES will
only lead to a world of pain.

Story/2423

Signed-off-by: Peter Clark <peter.clark@unipart.io>