[IMP] udes_stock: Implement delivery slip for multiple source documents
Add source document to each stock move line. Rename table headings and
remove source document reference and add picking reference number together
with date.

The customer address is shown when a partner is present, this is needed
for Check picking types but not for others. This issue will need to be
addressed in the future.

Story: x2422

Signed-off-by: Lorenzo Cucurachi <lorenzo.cucurachi@unipart.io>