[FIX] udes_stock_refactoring: Add refactor support for package levels
When refactoring a picking, ensure the package levels are setup as
needed on the new picking(s).

If this is not done, then there is a chance of Odoo creating a new
package level later via `_check_entire_pack`. This will have the
knock-on effect of `_generate_moves` being called when confirming
a picking, and means an extra unneeded move will then be created.

Story/2911

Signed-off-by: Peter Clark <peter.clark@unipart.com>