[FIX] udes_stock: Fix logic for checking if a picking needs a backorder
Previously validate_picking was only taking into account partially
completed move lines, and not incomplete moves.

This is important because a picking in the middle of a chain may
have been completed while stock was only partially available.
So the move line was completed, however the move was not
and had no move line to show this as stock had not yet been
assigned.

Story/2911

Signed-off-by: Peter Clark <peter.clark@unipart.com>