[FIX] udes_stock: Fix logic for update_orig_ids in Stock Move
Updated logic in update_orig_ids so that when a move is split from
being backordered, each move points to the correct earlier move.

Moves are first linked via move lines, and if a move doesn't have
a move line then it will be checked after via a product/quantity
key.

For example:

PICK1 for 10 apples completed 9, with a backorder
created for the last one.

OUT1 is then completed for these 9 apples, with a backorder made for the
remaining apple. When creating the backorder, the move for the 9 apples
needs to point to the previously completed Pick. The move for the last
apple will need to point to the backorder pick, so that when this is
completed the backorder goods out will be set to assigned.

Story/2911

Signed-off-by: Peter Clark <peter.clark@unipart.com>