[IMP] udes_stock: Reserve unbatched pickings in groups.
story/3625

Signed-off-by: Kevin Dwyer <kevin.dwyer@unipart.com>

The stock reservation job processes pickings one at a time, unless the
picking is in a batch, in which case all the pickings in the batch are
processed together. This works well for environments where batches are
used, as it allows us to take into account batch-level concerns, for
example atomic batch reservation; however it's inefficient if batches
are not used, as stock is reserved one picking at a time.

This patch separates the processing of batched and not-batched pickings,
allowing pickings that aren't members of batches to be processed in
groups rather than singly.

It has the pleasing side-effect of reducing the number of log messages
generated by the job in environments where batches are not used.