[IMP] udes_stock: Reduce stock reservation queries.
story/3625

Signed-off-by: Kevin Dwyer <kevin.dwyer@unipart.com>

Currently the stock reservation job will send a query for each confirmed
picking. The picking is then tested to see if it is in a batch, and if
the picking type requires atomic batch reservation then all the pickings
from the batch are fetched and processed.

This patch changes the processing so that we fetch pickings in chunks
and then loop over the pickings as before. This should improve the
efficency of the stock reservation job at the cost of making manual
stock reservation slightly slower as unbatched pickings that might have
been processed together will now be processed individually. It might be
possible to retain the old behaviour for manual reservation, but it
would increase code complexity, and I'd like to think that manual
reservations would consist of relatively few pickings.