[IMP] udes_stock: Reduce queries in reserve_stock
story/3625

Signed-off-by: Kevin Dwyer <kevin.dwyer@unipart.com>

Currently the stock reservation job will send a query for each confirmed
picking. The picking is then tested to see if it is in a batch, and if
the picking type requires atomic batch reservation then all the pickings
from the batch are fetched and processed.

This patch changes the processing so that we fetch pickings in chunks
and then loop over the pickings as before. This should improve the
efficency of the stock reservation job in cases where the number of
reservable pickings defined on the picking type is large or unlimited.