[FIX] udes_stock: Match fragmented original move ids.
story/3960

Signed-off-by: Kevin Dwyer <kevin.dwyer@unipart.com>

When matching moves with their original moves we match on product and
quantity. However this method will fail to match if a move relates to
more than one original move, for example if a move with quantity 6 is
related to previous moves of quantities 3, 2 and 1.

This patch attempts to handle this case by checking whether we have a
single move to match, and its quantity matches the total quantity of all
the candidate original moves. If this condition is true, the move's
`move_orig_ids` is set to all of the candidates.