[FIX] udes_stock: Ensure partner id is set on moves.
story/2419

Signed-off-by: Kevin Dwyer <kevin.dwyer@unipart.com>

If partner id is not set on moves, the core StockMove._assign_picking
will not correctly group pickings further down the chain correctly, and
consequently will clear pickings' origin and partner_id fields.

To prevent this behaviur, this patch sets partner_id on moves that are
created via the desktop.