[ADD] udes_stock: Add u_is_picking_zone flag.
story/4528

Signed-off-by: Kevin Dwyer <kevin.dwyer@unipart.com>

Add a new flag, `u_is_picking_zone` to the `StockLocation` model to
support refactoring by picking zone.

Update the location details view to display the flag for view locations,
and add a filter.

Add a new computed field, `u_picking_zone_id` to identify the picking
zone that a location inhabits. Display this field in the view if it
is set.

Signed-off-by: Kevin Dwyer <kevin.dwyer@unipart.com>