[IMP] udes_sale_stock: allow reducing quantity of sale orders in editable state
SET-143

Allow editing sale orders via EDI, if order is in editable state.
Order can be edited if it is not in progress, done or cancelled.
Order is considered 'in progress' if linked picking is done or
it is assigned and attached to a batch.

Stock move quantities are adjusted by Odoo if quantity is higher.
If quantity was lower in EDI file, manually set quantities, if
moves are still editable.

Signed-off-by: Rajeesh KV <rajeesh.kv@unipart.com>